requirements.txt
django==4.2
djangorestframework==3.15.2
pandas
matplotlib
djangorestframework-simplejwt
django-cors-headers
mysqlclient
